requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asyncio==3.4.3
attrs==25.3.0
cattrs==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
idna==3.10
magic-filter==1.0.12
mercadopago==2.3.0
multidict==6.4.4
platformdirs==4.3.8
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
requests-cache==1.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
url-normalize==2.2.1
urllib3==2.4.0
yarl==1.20.0
